refactor: abstract export component
